rename rbac resource from fluxauth  to automation
