+++ service/controller/rbac/resource/automation/create.go
@@ -1,4 +1,4 @@
-package fluxauth
+package automation
 
 import (
 	"context"
