+++ service/controller/rbac/resource/automation/delete.go
@@ -1,4 +1,4 @@
-package fluxauth
+package automation
 
 import (
 	"context"
