+++ service/controller/rbac/resource/automation/resource.go
@@ -1,11 +1,11 @@
-// fluxauth package is responsible for setting up RBAC required by
+// automation package is responsible for setting up RBAC required by
 // customer-facing Flux, operating in flux-system namespace.
 // This includes creating RoleBindings to upstream Flux roles in Organization
 // namespaces, RoleBindings allowing default Flux ServiceAccounts access to
 // said namespaces, as well as creating an `automation` ServiceAccount in every
 // Organization namespace, with enough permissions to reconcile Flux resources
 // there.
-package fluxauth
+package automation
 
 import (
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
@@ -16,7 +16,7 @@ import (
 )
 
 const (
-	Name = "fluxauth"
+	Name = "automation"
 )
 
 type Config struct {
